some code reorder
